+++ packages/at_auth/example/onboard.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 import 'package:args/args.dart';
+import 'package:at_chops/at_chops.dart' show SigningAlgoType;
 import 'package:at_auth/at_auth.dart';
 
 /// Perform initial onboarding for an atsign
@@ -24,7 +25,7 @@ void main(List<String> args) async {
 
     final atAuth = AtAuth.create();
     final atSign = argResults['atsign'];
-    final atOnboardingRequest = AtOnboardingRequest(atSign)
+    final atOnboardingRequest = AtOnboardingRequest(atSign, signingAlgoType: SigningAlgoType.rsa2048)
       ..rootDomain = argResults['rootDomain'];
     final atOnboardingResponse =
         await atAuth.onboard(atOnboardingRequest, argResults['cramsecret']);